pipes: simplify pipe capacity setting
When setting pipe capacities, we can get -EBUSY because the pipe may be
even larger than we thought, with data already in it. We can also
get -EPERM, because we don't have permission to increase the pipe
capacity. These two reasons are enough to simply stop caring about the
errors when setting pipe capacities.

Well, that's true except when doing load balancing on pipes, because we
need to know the capacity to be able to avoid blocking on writes.

Signed-off-by: Nicolas Viennot <Nicolas.Viennot@twosigma.com>